chore: extracted enums to their own files
