+++ GlucoseTray.Tests/DSL/Read/ReadDriver.cs
@@ -1,4 +1,5 @@
-﻿using GlucoseTray.Read.Dexcom;
+﻿using GlucoseTray.Enums;
+using GlucoseTray.Read.Dexcom;
 using GlucoseTray.Read.Nightscout;
 using NSubstitute;
 namespace GlucoseTray.Tests.DSL.Read;
