+++ GlucoseTray/Display/GlucoseDisplay.cs
@@ -1,4 +1,6 @@
-﻿namespace GlucoseTray.Display;
+﻿using GlucoseTray.Enums;
+
+namespace GlucoseTray.Display;
 
 public class GlucoseDisplay
 {
@@ -14,13 +16,4 @@ public class GlucoseDisplay
         var staleMessage = IsStale ? $"\r\n{Math.Abs((utcNow - TimestampUtc).TotalMinutes):#} minutes ago" : string.Empty;
         return $"{DisplayValue} {TimestampUtc.ToLocalTime().ToLongTimeString()} {Trend.GetTrendArrow()} {staleMessage}".Trim();
     }
-}
-
-public enum IconTextColor
-{
-    White,
-    Black,
-    Yellow,
-    Gold,
-    Red,
 }
