+++ GlucoseTray/Enums/AlertLevel.cs
@@ -1,0 +1,10 @@
+﻿namespace GlucoseTray.Enums;
+
+public enum AlertLevel
+{
+    None,
+    CriticalLow,
+    Low,
+    High,
+    CriticalHigh,
+}
