+++ GlucoseTray/Enums/DexcomServer.cs
@@ -1,0 +1,8 @@
+﻿namespace GlucoseTray.Enums;
+
+public enum DexcomServer
+{
+    DexcomShare1,
+    DexcomShare2,
+    DexcomInternational
+}
