+++ GlucoseTray/Enums/GlucoseSource.cs
@@ -1,0 +1,7 @@
+﻿namespace GlucoseTray.Enums;
+
+public enum GlucoseSource
+{
+    Dexcom,
+    Nightscout,
+}
