+++ GlucoseTray/Enums/GlucoseUnitType.cs
@@ -1,0 +1,7 @@
+﻿namespace GlucoseTray.Enums;
+
+public enum GlucoseUnitType
+{
+    Mg,
+    Mmol
+}
