+++ GlucoseTray/Enums/IconTextColor.cs
@@ -1,0 +1,10 @@
+﻿namespace GlucoseTray.Enums;
+
+public enum IconTextColor
+{
+    White,
+    Black,
+    Yellow,
+    Gold,
+    Red,
+}
