+++ GlucoseTray/appsettings.json
@@ -1,4 +1,7 @@
 ﻿{
+  "appsettings": {
+    "Version": "15.2.3"
+  },
   "IsDarkMode": false,
   "MinutesUntilStale": 0,
   "RefreshIntervalInMinutes": 0,
